Revert "Condense login form"
This reverts commit a1159e0016fa00bb11e34e052b603584b345e74f.

That change might be looking cool on windows, but on other OS it's now totaly borked.
You need to use auto-sizing layouts if you really want to redesign it. Or tell me how do you
want it to look and I will do that, but hardcoding positions and sizes will never
work on all platforms, the fonts and sizes are not same everywhere.

Try this on gnome :P