fixed block user
There was a problem with dependency handling, we modified the dependency status after it
was changed to "IsProcessed()" while not being marked as failed, and then we additionaly marked
it as failed in separate thread. This caused race condition and sometimes could result in sending
message even if block has failed. Now the message is sent only after the block is successfuly
finished so it can't happen.