fixed
There was a stupid bug in code that inserted and edit to queue, if edit
had too low score it was inserted using append insted of putting it to
position max - 1. That is necessary because last item is not a queue item
but a spacer that keep the items on top. This bug was in huggle for long
time but it was triggered only now because thanks to edit count, it was
finally possible to create edits with very low edit score, even lower than
a minimal score is.