token
during edit if we get a badtoken error we can try to invalidate the token
in config and get a new one. This doesn't really save the current edit
query, but may fix similar problems in future without having to restart
huggle, which is now probably only workaround to fix it.