huggleweb.cpp: Use red colour for summaries
This commit changes the colour of summaries to red, and blank summary placeholder to black.
Patrollers sometimes go too fast when patrolling, ignoring edit summaries which may justify the blanking/suspicious edits.
In conclusion, we need to notify the patrollers when an edit /has/ a summary, not when it doesn't.