no more instant reverts by default
Now the instant reverts need to be explicitly enabled in configuration,
this makes it possible to cancel reverts by default. It will make huggle
slower by 6 seconds per revert but it might help lower down the amount of
reverts that were caused by misclicks and so on.

Basically now when you hit the revert button, huggle will wait for few
seconds before it actually performs the revert. We had this feature all
time, but it wasn't enabled by default.

This will affect new huggle installations only.