Resolved network issue
We used to store a temporary part of network result of every ApiQuery
in a string and on every ReadData event we converted that QByteArray
into a string and appended it to this temporary string.

This was wrong as the QByteArray didn't contain whole Unicode text but
only a part of it that might be just a half of a symbol or just something
invalid that couldn't be parsed by QString.

That resulted in randomly corrupt symbols in some pages, that in case of
software rollback were written back to wikipedia.

Now we store the parts of QByteArray to a temporary QByteArray which
converted to QString in one shot later. It uses a bit more operating
memory and less CPU, but it works.