Contribution browser
We have all the code necessary to display information about contributions in ReportForm. So this commit basically makes it possible to turn it into a contribution browser which can be opened from anywhere else in huggle. Thanks to this we no longer need to offer contributions only as external link to wiki but we can also display them locally inside of huggle.

This also removes one hardcoded shortcut for a feature that isn't really widely used.