Localized some strings
Localized some strings in login.cpp and sessionform.cpp. No changes were made
to the content of the strings, though some grammatical touch ups may be of
use in login.cpp. To the people at GCI: the 20 changes required were not
accomplished due to personal reasons, however, I expect to be making a total
of 20 changes before Dec 24 2015.

Bug: T67950