Fixed some issues with previous commits
This changes the behaviour of HAN a bit, but adds a lot of performance. Now we are using more standard-for-chats scrolling, where text is added to bottom instead of top of the text box. We are using Qt native calls for that instead of our own.

There are 2 issues with this:
- Plain text doesn't support links, I made a nasty hack for it to work.
- We have no control over the current buffer, so there may need to be an option to clear it, to prevent memory hogging (on other hand the text in chat window doesn't really take much ram)