implemented auto-refresh
So that if we are looking on some edit and there was a new edit done to
that very page we are looking at, it automatically reloads to top
revision. This is exactly how it worked in Huggle 2x and we kinda forgot
to implement this. It's optional but enabled by default.