shortcut hash check
This is a safety check on each login to huggle to verify whether warning
menus didn't change. Right now we store the shortcuts in a way that we
set shortcut for Nth item in the list. In case someone changes the
project configuration, Nth item may no longer correspond to what it was
and shortcut would start triggering different templates. This patch
implements a MD5 sum made from warning menu that is compared every login
and in case there is a difference, shortcuts for menus will be reset
and user is warned.