try to find system yaml-cpp before installing submodule
Huggle packages have been broken for a while on any linux distro that
also packages yaml-cpp. This has lead to vairous packaging hacks that
don't work reliably.

This patch tells CMake to try to find the system yaml-cpp package first.
If the package is found, the include directory is included.
If the package is not found, the existing routine to install yaml-cpp
from the git submodule is run.

Huggle builds and runs using system yaml-cpp on my system with this
patch, but it does not without the patch. I didn't test it without
yaml-cpp installed because I have another package that depends on it
that I don't want to uninstall.

Fixes #277