Remove deprecated Windows Server 2016
This PR removes the now deprecated Windows Server 2016 Github Actions runner image, and should fix the currently failing CI. This is a bit of a band-aid fix, as the correct solution is to ensure that Huggle builds on Server 2022 and the latest Visual Studio, which I have tested and does not currently work. This will fix #355.