QT6 support
We need to move on to latest supported version of Qt. With this change I
would like to deprecate Qt4 completely and stay with support for only
Qt5 and Qt6.

There are still many code artifacts reflecting Qt4 and old WebKit, which
we can gradually get rid of now.

WebKit is probably supported by early Qt5 but I would like to phase it
out as well.