generic code cleanup
* Implemented function to reset network manager
We don't use it (yet) but this could be used to tackle some problems with
expired tokens and failed attempts to relogin. Problem is that right now
we have single network manager for entire Huggle and share it for all
wikis. This needs to change so that we can actually take advantage of such
reset, or else we would need to logout from all projects, which isn't
efficient.

* Moved some inline definition out to .cpp, there is no point in that
  being inline, it's barely used

* Used AI copilot to rename many UI elements that didn't have meaningful name